Not Finished HealthSaver

+++ Assets/Scripts/HP/DamageDealer.cs
@@ -22,7 +22,7 @@ public class DamageDealer : MonoBehaviour
             {
                 if (hit.transform != null && hit.transform.TryGetComponent<IDamageable>(out IDamageable damageable))
                 {
-                    damageable.TakeDamage(damage, _damageTo);
+                    damageable.TakeDamage((int)damage, _damageTo);
 #if(UNITY_EDITOR)
                     Debug.Log("hit " + hit.transform.name);
 #endif
