+++ Assets/Scripts/HP/IDamageable.cs
@@ -2,7 +2,7 @@ using UnityEngine;
 
 public interface IDamageable 
 {
-    public float Hp { get; set; }
-    void TakeDamage(float damage, LayerMask mask);
+    public int Hp { get; set; }
+    void TakeDamage(int damage, LayerMask mask);
     void CheckDeath();
 }
